src/requirements.txt
PyGitHub==1.55
pygit2==1.6.0
kubernetes==17.17.0